ver 0.1.1

+++ employee/src/main/java/com/ibm/wude/controller/EmployeeController.java
@@ -2,10 +2,10 @@ package com.ibm.wude.controller;
 
 import java.util.List;
 
+import org.apache.ibatis.annotations.Param;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.ibm.wude.model.EmployeeModel;
@@ -35,6 +35,17 @@ public class EmployeeController {
 	}
 
 	/**
+	 * 查询员工信息
+	 * 
+	 * @param name
+	 * @return
+	 */
+	@GetMapping("/getEmpByName/{name}")
+	public EmployeeModel getEmployeeModelByName(@PathVariable("name") String name) {
+		return employeeService.getEmployeeByName(name);
+	}
+
+	/**
 	 * 删除员工信息
 	 * 
 	 * @param id
@@ -43,8 +54,14 @@ public class EmployeeController {
 	// @DeleteMapping("/delEmpById/{id}")
 	@GetMapping("/delEmpById/{id}")
 	public boolean deleteEmployeeById(@PathVariable("id") Integer id) {
-		System.out.println("id:" + id);
-		return employeeService.deleteEmployeeById(id);
+		EmployeeModel emp = new EmployeeModel();
+		emp.setId(id);
+		if (getEmployeeModelById(id) != null) {
+			return employeeService.deleteEmployeeById(id);
+		} else {
+			return false;
+		}
+
 	}
 
 	/**
@@ -55,7 +72,17 @@ public class EmployeeController {
 	 */
 	// @PutMapping("/updateEmpById")
 	@GetMapping("/updateEmp")
-	public boolean updateEmployee(@RequestBody EmployeeModel employeeModel) {
-		return employeeService.updateEmployee(employeeModel);
+	public boolean updateEmployee(@Param("id") Integer id, @Param("name") String name, @Param("salary") String salary,
+			@Param("age") Integer age) {
+		EmployeeModel emp = new EmployeeModel();
+		emp.setId(id);
+		emp.setName(name);
+		emp.setSalary(salary);
+		emp.setAge(age);
+		if (getEmployeeModelById(id) != null) {
+			return employeeService.updateEmployee(emp);
+		} else {
+			return false;
+		}
 	}
 }
